refactor(bevy-migration): 推进v3.1迁移进度，完成多项核心更新
本次提交完成大量Bevy 0.19迁移工作：
1.  为多个核心组件、资源添加Reflect支持：TagSet/AggregatorState/BattleUnitId/FrameCounter/AutoSaveConfig/MetricsCollector
2.  修复cutscene系统的Timer到Delayed Commands迁移
3.  更新迁移文档至v3.1版本，更新完成度和剩余工作清单
4.  补充SceneRoot模块导入，更新部分文档状态和内容

+++ src/core/domains/narrative/systems/cutscene_system.rs
@@ -35,7 +35,7 @@ pub(crate) fn on_cutscene_start(trigger: On<CutsceneStartRequest>, mut commands:
     // 使用 Delayed Commands 替代逐帧 tick：duration 秒后自动触发结束
     commands
         .delayed()
-        .secs(req.duration as f64)
+        .secs(req.duration)
         .trigger(CutsceneEnded {
             cutscene_id: req.cutscene_id.clone(),
         });
